Add function grow()
This function automatically expands the hashmap if the total number of
elements if equal or higher than the set limit, which in our case
is capacity * 0.75.
Using the length() and entries() functions we check if the
buckets' length has been matched, then we copy all the entries
into a temporary map and back into the current one with expanded
capacity.